+++ agilepy/core/CustomExceptions.py
@@ -118,4 +118,8 @@ class SourcesLibraryIsEmpty(Exception):
     def __init__(self, message):
         super().__init__(message)
 
+class ValueOutOfRange(Exception):
+    def __init__(self, message):
+        super().__init__(message)
+